The arity table for protocols now starts at 1
The self parameter used to be counted in the arity of a method
so that all methods would habe an arity of one. This has been
changed so that the self parameter doesn't count towards the arity.
A method with one parameter is now put in the arity table
at position 0.